Rename operator -> signer in tokens
Within the tokens plugin, the term "operator" was used to indicate the
transaction signer. This could be confusing, because within EIP-721 &
EIP-1155 operator is an address authorized to perform token transfers.

Since FireFly is planning on providing support for token approvals,
operator as it was previously defined is now known as "signer". Now
FireFly can reintroduce operator in the approval context.

Signed-off-by: Alex Shorsher <alex.shorsher@kaleido.io>